Major fixes and fixing broken events
Basicaly, everything was broken, there were no permissions and the
events weren't even registered!
All of that is now fixed in this new version and I added a few new
features to the signs such as, it tells you how many signs are currently
registered.
All of the data is stored in the config.yml for now until I make it
data.yml.